Rename MapViewModule.ts
